seperated into kitefuncs-file
